+++ react-components/src/components/Signup/Form.jsx
@@ -25,9 +25,9 @@ const Form = ({
     className="signup__form"
   >
     {showTitle && (
-      <legend className="h-s text-blue-deep-80 p-t-xs">
+      <h1 className="h-s text-blue-deep-80 p-t-xs">
         Create an account
-      </legend>
+      </h1>
     )}
     Already have an account? &nbsp;&nbsp;
     <a