+++ react-components/src/components/ToggleSnapshot/index.jsx
@@ -15,6 +15,8 @@ export const ToggleSnapshot = memo(({
           className='button button--tertiary button--icon'
           type='button'
           onClick={() => setToggle(!toggle)}
+          aria-controls="data-snapshot"
+          aria-expanded={toggle}
         >
           <i className='fas fa-chart-bar' />{`${toggle ? 'Hide' : 'Open'} Data Snapshot`}
         </button>