+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Enhancements
 - NOTICKET - update changelog, post-release
 ### Fixed bugs
+- GP2-2884 - DAC Focus Order 01
+- GP2-2893 - DAC Link Purpose 01
 - GP2-2890 - DAC NonText Content 01
 - GP2-2835 - Disable caching on homepage
 - GP2-2889 - DAC Decorative_Images_01 - clear alt tags on decorative images