+++ CHANGELOG.md
@@ -18,6 +18,10 @@
 - GP2-2907 - DAC_Headings_06
 - GP2-2929 - DAC_Accordions_01
 - GP2-2931 - DAC_Accordions_03
+- GP2-2916 - DAC_CSS_Content_01
+- GP2-2917 - DAC_CSS_Content_02
+- GP2-2918 - DAC_CSS_Content_03
+- GP2-2919 -DAC_CSS_Content_04
 
 ## [1.10.1](https://github.com/uktrade/great-cms/releases/tag/1.10.1)
 ### Enhancements